+++ backend/api/gamemaps/gamemap/__mocks__/gamemap.queries.js
@@ -1,0 +1,8 @@
+const mockGameMapQueries = {
+    getTopScoresForGameMap: jest.fn(),
+    updateGameMapDetails: jest.fn().mockResolvedValue(true),
+    getAllImageIdsForGameMap: jest.fn(),
+    deleteGameMapById: jest.fn().mockResolvedValue(true)
+};
+
+module.exports = mockGameMapQueries;
